Add script to send files to remote server
This commit adds a script to send files to a remote server. The script uses SSH to securely transfer files from the local directory "./para" to the remote server's directory "$REMOTE_BASE_DIR/matrix". The script is triggered by finding all "*.npy" files in the local directory and executing the necessary commands. Additionally, the commit includes minor changes to the script's execution order.